Use nightly JDK releases

+++ client.ps1
@@ -27,14 +27,11 @@ $ProgressPreference = 'SilentlyContinue'
 $ErrorActionPreference = 'Stop'
 $PSScriptRoot = Split-Path $MyInvocation.MyCommand.Path -Parent
 
+$_java_version = 11
 if ($env:PROCESSOR_ARCHITECTURE -eq 'AMD64' -or $env:PROCESSOR_ARCHITEW6432 -eq 'AMD64') {
 	$_arch = 'x64'
-	$_java_version = 11
-	$_java_type = 'jdk'
 } else {
 	$_arch = 'x32'
-	$_java_version = 8
-	$_java_type = 'jre'
 }
 
 $_jre_dir = Join-Path $PSScriptRoot jre-$_java_version-windows-$_arch
@@ -43,30 +40,29 @@ if (!(Test-Path $_jre_dir)) {
 	$_temp_dir = Join-Path $PSScriptRoot temp
 	New-Item -ItemType Directory -Path $_temp_dir -Force -Verbose | Out-Null
 	
-	$_temp_jdk_archive = Join-Path $_temp_dir $_java_type-$_java_version-windows-$_arch.zip
+	$_temp_jdk_archive = Join-Path $_temp_dir jdk-$_java_version-windows-$_arch.zip
 	
 	if (!(Test-Path $_temp_jdk_archive)) {
-		Download-File -Uri "https://api.adoptopenjdk.net/v2/binary/releases/openjdk$($_java_version)?openjdk_impl=hotspot&release=latest&type=$_java_type&heap_size=normal&os=windows&arch=$_arch" -OutFile $_temp_jdk_archive
+		Download-File -Uri "https://api.adoptopenjdk.net/v2/binary/nightly/openjdk$($_java_version)?openjdk_impl=hotspot&release=latest&type=jdk&heap_size=normal&os=windows&arch=$_arch" -OutFile $_temp_jdk_archive
 	}
 	
-	$_temp_jdk_dir = Join-Path $_temp_dir $_java_type-$_java_version-windows-$_arch
+	$_temp_jdk_dir = Join-Path $_temp_dir jdk-$_java_version-windows-$_arch
 	New-Item -ItemType Directory -Path $_temp_jdk_dir -Force -Verbose | Out-Null
 	
 	Expand-Zip -Path $_temp_jdk_archive -DestinationPath $_temp_jdk_dir
 	
 	$_jdk_home = Join-Path $_temp_jdk_dir * -Resolve
-	if ($_java_type -eq 'jdk') {
-		& "$_jdk_home\bin\jlink" -v `
-		 --no-header-files `
-		 --no-man-pages `
-		 --strip-debug `
-		 --compress=2 `
-		 --module-path "$_jdk_home\jmods" `
-		 --add-modules java.desktop,java.management `
-		 --output "$_jre_dir"
-	} else {
-		Copy-Item $_jdk_home -Destination $_jre_dir -Recurse
-	}
+	
+	& "$_jdk_home\bin\jlink" -v `
+	 --no-header-files `
+	 --no-man-pages `
+	 --strip-debug `
+	 --compress=1 `
+	 --module-path "$_jdk_home\jmods" `
+	 --add-modules java.desktop,java.management `
+	 --output "$_jre_dir"
+	
+	if ($LastExitCode -ne 0) { exit $LastExitCode }
 	
 	Remove-Item $_temp_dir -Recurse
 }
@@ -82,5 +78,7 @@ New-Item -ItemType Directory -Path $_cache_dir -Force -Verbose | Out-Null
  "-Dsun.awt.noerasebackground=true" `
  "-Dcom.jagex.configuri=jagex-jav://oldschool.runescape.com/jav_config.ws" `
  "$_jar" "$((Get-Item $PSScriptRoot).Name)"
+
+if ($LastExitCode -ne 0) { exit $LastExitCode }
  
 Remove-Item $_jar -Verbose
